Tidied up existing test classes and added InfectionLogTest class
