+++ src/test/java/uk/co/ramp/covid/simulation/covid/CovidTest.java
@@ -1,7 +1,8 @@
-package uk.co.ramp.covid.simulation;
+package uk.co.ramp.covid.simulation.covid;
 
 import org.junit.After;
 import org.junit.Test;
+import uk.co.ramp.covid.simulation.Time;
 import uk.co.ramp.covid.simulation.covid.Covid;
 import uk.co.ramp.covid.simulation.covid.CovidParameters;
 import uk.co.ramp.covid.simulation.place.Household;
