+++ src/test/java/uk/co/ramp/covid/simulation/place/ConstructionSiteTest.java
@@ -66,11 +66,8 @@ public class ConstructionSiteTest extends SimulationTest {
     @Test
     public void testConstructionSiteWorkers() throws ImpossibleWorkerDistributionException {
         int populationSize = 10000;
-        int nInfections = 10;
-
         Population p = PopulationGenerator.genValidPopulation(populationSize);
         p.allocatePeople();
-        p.seedVirus(nInfections);
         List<Person> staff;
         Time t = new Time(0);
         //Run for a whole week
