+++ src/test/java/uk/co/ramp/covid/simulation/place/RestaurantTest.java
@@ -66,11 +66,8 @@ public class RestaurantTest extends SimulationTest {
     @Test
     public void testRestaurantWorkers() throws ImpossibleWorkerDistributionException {
         int populationSize = 10000;
-        int nInfections = 10;
-
         Population p = PopulationGenerator.genValidPopulation(populationSize);
         p.allocatePeople();
-        p.seedVirus(nInfections);
         List<Person> staff;
         Time t = new Time(0);
         //Run for a whole week
