+++ src/test/java/uk/co/ramp/covid/simulation/place/SchoolTest.java
@@ -26,11 +26,8 @@ public class SchoolTest extends SimulationTest {
     @Test
     public void testSchoolWorkers() throws ImpossibleWorkerDistributionException {
         int populationSize = 10000;
-        int nInfections = 10;
-
         Population p = PopulationGenerator.genValidPopulation(populationSize);
         p.allocatePeople();
-        p.seedVirus(nInfections);
         List<Person> staff;
         Time t = new Time(0);
         //Run for a whole week
