+++ src/test/java/uk/co/ramp/covid/simulation/place/ShopTest.java
@@ -65,11 +65,8 @@ public class ShopTest extends SimulationTest {
     @Test
     public void testShopWorkers() throws ImpossibleWorkerDistributionException {
         int populationSize = 10000;
-        int nInfections = 10;
-
         Population p = PopulationGenerator.genValidPopulation(populationSize);
         p.allocatePeople();
-        p.seedVirus(nInfections);
         List<Person> staff;
         Time t = new Time(0);
         //Run for a whole week
