+++ src/test/java/uk/co/ramp/covid/simulation/population/PopulationTest.java
@@ -3,7 +3,6 @@ package uk.co.ramp.covid.simulation.population;
 import org.junit.Before;
 import org.junit.Test;
 import uk.co.ramp.covid.simulation.DailyStats;
-import uk.co.ramp.covid.simulation.RStats;
 import uk.co.ramp.covid.simulation.place.*;
 import uk.co.ramp.covid.simulation.place.householdtypes.*;
 import uk.co.ramp.covid.simulation.testutil.PopulationGenerator;
@@ -351,36 +350,4 @@ public class PopulationTest extends SimulationTest {
 
         assertTrue(child > infant);
     }
-
-    @Test
-    public void secondaryInfectionsAreLogged() {
-        pop.seedVirus(10);
-        pop.simulate(20);
-        int totalSecondary = 0;
-        for (Person p : pop.getAllPeople()) {
-            if (p.getcVirus() != null) {
-                totalSecondary += p.getcVirus().getInfectionLog().getSecondaryInfections().size();
-            }
-        }
-        assertTrue(totalSecondary > 0);
-    }
-
-    @Test
-    public void symptomaticCasesAreLogged() {
-        pop.seedVirus(1);
-
-        Person infected = null;
-        for (Person p : pop.getAllPeople()) {
-            if (p.getcVirus() != null) {
-                infected = p;
-            }
-        }
-        
-        infected.getcVirus().forceSymptomatic(true);
-        
-        pop.simulate(50);
-        
-        assertNotNull(infected.getcVirus().getInfectionLog().getSymptomaticTime());
-    }
-
 }
